+++ tests/integration/models/test_pickle_models.py
@@ -1,5 +1,6 @@
-import pytest
 from typing import Any
+
+import pytest
 
 from tests.models.pickle_types import ModelWithUnserializableFields